backend/pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = "FastAPI boilerplate backend service"
authors = [{ name = "Your Name", email = "you@example.com" }]
dependencies = [
    "fastapi>=0.111.0,<0.112.0",
    "uvicorn[standard]>=0.30.0,<0.31.0",
    "pydantic-settings>=2.2.1,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "sqlalchemy>=2.0,<3.0",
    "aiosqlite>=0.18,<0.19",
]
[project.optional-dependencies]
dev = [
    "httpx>=0.27.0,<0.28.0",
    "pytest>=8.2.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<0.24.0",
    "ruff>=0.5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]
